Polish Pass: Part 1 (#4864)

+++ Kavita.Models/DTOs/UpdateVolumeDto.cs
@@ -11,6 +11,6 @@ public sealed record UpdateVolumeDto : IUpdateExternalMetadataIds
     public int? HardcoverId { get; set; }
     public long? MetronId { get; set; }
     public string ComicVineId { get; set; }
-    public long? MangaBakaId { get; set; }
+    public int? MangaBakaId { get; set; }
     public int? CbrId { get; set; }
 }
